requirements_api.txt
# API V2 - Production Requirements
# =================================

# FastAPI Framework
fastapi[all]==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2

# Rate Limiting
slowapi==0.1.9

# Database
psycopg2-binary==2.9.9
SQLAlchemy==2.0.25

# Configuration
python-dotenv==1.0.0
pydantic-settings==2.1.0

# PDF Processing (existentes)
PyMuPDF==1.23.8
PyPDF2==3.0.1

# Logging y Monitoring
python-json-logger==2.0.7

# Testing (opcional)
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# CORS
python-cors==1.0.0
